debug

+++ Shell/Class/Functions/CatFunction.cs
@@ -15,6 +15,7 @@ namespace Shell.Class.Functions
         static CultureInfo culture = new CultureInfo("fr-FR");
 
         static string pathSource;
+        static string pathSource2;
         static string pathDest;
         static string creatDateV;
         static string modifDateV;
@@ -24,6 +25,7 @@ namespace Shell.Class.Functions
         {
             command = Main.Command;
             pathSource = "";
+            pathSource2 = "";
             pathDest = "";
             creatDateV = "";
             modifDateV = "";
@@ -40,27 +42,68 @@ namespace Shell.Class.Functions
 
             if (File.Exists(pathSource))
             {
-                if (pathSource != "" && pathDest != "")
+                if (File.Exists(pathSource2))
                 {
-                    if (Directory.Exists(pathDest))
+                    if (pathSource != "" && pathSource2 != "" && pathDest != "")
                     {
-                        File.Copy(pathSource, pathDest + "/" + Path.GetFileName(pathSource), true);
+                        if (!Directory.Exists(pathDest))
+                        {
+                            string res = "";
+
+                            foreach (var item in File.ReadAllLines(pathSource))
+                            {
+                                res += item + Environment.NewLine;
+                            }
+                            foreach (var item in File.ReadAllLines(pathSource2))
+                            {
+                                res += item + Environment.NewLine;
+                            }
+
+                            File.WriteAllText(pathDest, res);
+                        }
+                        else
+                        {
+                            Console.WriteLine("Dossier portant le même nom déjà existant.");
+                            return 0;
+                        }
                     }
-                    else if (pathDest.Replace("/", "\\")[pathDest.Length - 1] == '\\')
+                    else if (creatDateV == "" && modifDateV == "")
                     {
-                        Directory.CreateDirectory(pathDest);
-                        File.Copy(pathSource, pathDest + "/" + Path.GetFileName(pathSource), true);
-                    }
-                    else
-                    {
-                        File.Copy(pathSource, pathDest, true);
+                        foreach (var item in File.ReadAllLines(pathSource))
+                        {
+                            Console.WriteLine(item);
+                        }
                     }
                 }
-                else if (creatDateV == "" && modifDateV == "")
+                else
                 {
-                    foreach (var item in File.ReadAllLines(pathSource))
+                    if (pathSource != "" && pathDest != "")
                     {
-                        Console.WriteLine(item);
+                        if (Directory.Exists(pathDest))
+                        {
+                            File.Copy(pathSource, pathDest + "/" + Path.GetFileName(pathSource), true);
+                        }
+                        else if (pathDest.Replace("/", "\\")[pathDest.Length - 1] == '\\')
+                        {
+                            Directory.CreateDirectory(pathDest);
+                            File.Copy(pathSource, pathDest + "/" + Path.GetFileName(pathSource), true);
+                        }
+                        else
+                        {
+                            string dir = Path.GetDirectoryName(pathDest);
+                            if (!Directory.Exists(dir))
+                            {
+                                Directory.CreateDirectory(dir);
+                            }
+                            File.Copy(pathSource, pathDest, true);
+                        }
+                    }
+                    else if (creatDateV == "" && modifDateV == "")
+                    {
+                        foreach (var item in File.ReadAllLines(pathSource))
+                        {
+                            Console.WriteLine(item);
+                        }
                     }
                 }
             }
@@ -92,34 +135,29 @@ namespace Shell.Class.Functions
                 pathSource = DirectoryTool.SetPath(command.GetBaseValue(index).Replace("\"", ""));
                 pathDest = DirectoryTool.SetPath(command.GetBaseValue(index + 2).Replace("\"", ""));
                 index += 3;
-                if (command.IsCommandLike(index, "-m $value") || command.IsCommandLike(index, "--modifdate $value"))
+                if (IsDateModif() == 0)
                 {
-                    modifDateV = command.GetBaseValue(index + 1).Replace("\"", "");
-                    index += 2;
-                    if (command.IsCommandLike(index, "-c $value") || command.IsCommandLike(index, "--createdate $value"))
-                    {
-                        creatDateV = command.GetBaseValue(index + 1).Replace("\"", "");
-                    }
-                    else if (!command.IsCommandLike(index, "$end"))
-                    {
-                        return 0;
-                    }
-                }
-                else if (command.IsCommandLike(index, "-c $value") || command.IsCommandLike(index, "--createdate $value"))
-                {
-                    creatDateV = command.GetBaseValue(index + 1).Replace("\"", "");
-                    index += 2;
-                    if (command.IsCommandLike(index, "-m $value") || command.IsCommandLike(index, "--modifdate $value"))
-                    {
-                        modifDateV = command.GetBaseValue(index + 1).Replace("\"", "");
-                    }
-                    else if (!command.IsCommandLike(index, "$end"))
-                    {
-                        return 0;
-                    }
+                    return 0;
                 }
                 else if (!command.IsCommandLike(index, "$end"))
                 {
+                    Console.WriteLine("c");
+                    return 0;
+                }
+            }
+            else if(command.IsCommandLike(index, "$value $value > $value"))
+            {
+                pathSource = DirectoryTool.SetPath(command.GetBaseValue(index).Replace("\"", ""));
+                pathSource2 = DirectoryTool.SetPath(command.GetBaseValue(index + 1).Replace("\"", ""));
+                pathDest = DirectoryTool.SetPath(command.GetBaseValue(index + 3).Replace("\"", ""));
+                index += 4;
+                if (IsDateModif() == 0)
+                {
+                    return 0;
+                }
+                else if (!command.IsCommandLike(index, "$end"))
+                {
+                    Console.WriteLine("c");
                     return 0;
                 }
             }
@@ -179,7 +217,44 @@ namespace Shell.Class.Functions
 
                 return 0;
             }
+            else if (!command.IsCommandLike(index, "$end"))
+            {
+                return 0;
+            }
 
+            return 1;
+        }
+
+        static int IsDateModif()
+        {
+            if (command.IsCommandLike(index, "-m $value") || command.IsCommandLike(index, "--modifdate $value"))
+            {
+                modifDateV = command.GetBaseValue(index + 1).Replace("\"", "");
+                index += 2;
+                if (command.IsCommandLike(index, "-c $value") || command.IsCommandLike(index, "--createdate $value"))
+                {
+                    creatDateV = command.GetBaseValue(index + 1).Replace("\"", "");
+                }
+                else if (!command.IsCommandLike(index, "$end"))
+                {
+                    Console.WriteLine("a");
+                    return 0;
+                }
+            }
+            else if (command.IsCommandLike(index, "-c $value") || command.IsCommandLike(index, "--createdate $value"))
+            {
+                creatDateV = command.GetBaseValue(index + 1).Replace("\"", "");
+                index += 2;
+                if (command.IsCommandLike(index, "-m $value") || command.IsCommandLike(index, "--modifdate $value"))
+                {
+                    modifDateV = command.GetBaseValue(index + 1).Replace("\"", "");
+                }
+                else if (!command.IsCommandLike(index, "$end"))
+                {
+                    Console.WriteLine("b");
+                    return 0;
+                }
+            }
             return 1;
         }
     }
